checking in run_soundcast.py in the format it is most typically run NOW
in the future, many parts of this code will be uncommented.  this is how
we usually test for now, because we are not usually doing things like
copying the inputs over since the inputs are staying the same.
we are not building networks either.  for the most part, we are only running
daysim, the truck model, and skims and paths in a loop and testing changes.
several functions are left in because we would like to use them sometimes,
and also in the future